Split key mapper and frozen check from model
