+++ serious/serialization/key_mapper.py
@@ -1,0 +1,21 @@
+from abc import ABC, abstractmethod
+
+
+class KeyMapper(ABC):
+
+    @abstractmethod
+    def to_model(self, item: str) -> str:
+        raise NotImplementedError
+
+    @abstractmethod
+    def to_serialized(self, item: str) -> str:
+        raise NotImplementedError
+
+
+class NoopKeyMapper(KeyMapper):
+
+    def to_model(self, item: str) -> str:
+        return item
+
+    def to_serialized(self, item: str) -> str:
+        return item
